requirements.txt
streamlit
google-generativeai
yfinance
pandas
plotly
finance-data-reader
